+++ neo/Prompt/Commands/Invoke.py
@@ -24,6 +24,7 @@ from neo.SmartContract.ApplicationEngine import ApplicationEngine
 from neo.SmartContract import TriggerType
 from neo.SmartContract.StateMachine import StateMachine
 from neo.SmartContract.ContractParameterContext import ContractParametersContext
+from neo.SmartContract.ContractParameter import ContractParameterType
 from neo.SmartContract.Contract import Contract
 from neo.Core.Cryptography.Helper import scripthash_to_address
 from neo.Core.Cryptography.Crypto import Crypto
@@ -161,19 +162,17 @@ def InvokeWithTokenVerificationScript(wallet, tx, token, fee=Fixed8.Zero(), invo
     return False
 
 
-def TestInvokeContract(wallet, args, withdrawal_tx=None, from_addr=None,
-                       min_fee=DEFAULT_MIN_FEE, invoke_attrs=None, owners=None):
+def TestInvokeContract(wallet, args, withdrawal_tx=None, from_addr=None, min_fee=DEFAULT_MIN_FEE,
+                       invoke_attrs=None, owners=None, user_entry=False):
     BC = GetBlockchain()
 
     contract = BC.GetContract(args[0])
 
     if contract:
-        #
         params = args[1:] if len(args) > 1 else []
 
         params, neo_to_attach, gas_to_attach = PromptUtils.get_asset_attachments(params)
         params, parse_addresses = PromptUtils.get_parse_addresses(params)
-        params.reverse()
 
         if '--i' in params:
             params = []
@@ -182,6 +181,28 @@ def TestInvokeContract(wallet, args, withdrawal_tx=None, from_addr=None,
                 if abort:
                     return None, None, None, None, False
                 params.append(param)
+            params.reverse()
+        elif user_entry:
+            try:
+                i_args = []
+                for index, iarg in enumerate(contract.Code.ParameterList):
+                    ptype = ContractParameterType(iarg)
+                    param, abort = PromptUtils.verify_params(ptype, params[index])
+                    if abort:
+                        return None, None, None, None, False
+                    i_args.append(param)
+                i_args.reverse()
+                params = i_args
+            except IndexError:
+                print(f"Check inputs. {len(contract.Code.ParameterList)} params specified and only {len(params)} given.")
+                return None, None, None, None, False
+            except ValueError as e:
+                print("Check params.", e)
+                return None, None, None, None, False
+            except Exception as e:
+                print(f'Could not parse {params[index]} as {ptype}:', e)
+                return None, None, None, None, False
+        else:
             params.reverse()
 
         sb = ScriptBuilder()
@@ -363,7 +384,7 @@ def test_invoke(script, wallet, outputs, withdrawal_tx=None,
 
 def test_deploy_and_invoke(deploy_script, invoke_args, wallet,
                            from_addr=None, min_fee=DEFAULT_MIN_FEE, invocation_test_mode=True,
-                           debug_map=None, invoke_attrs=None, owners=None, enable_debugger=False, snapshot=None):
+                           debug_map=None, invoke_attrs=None, owners=None, enable_debugger=False, snapshot=None, user_entry=False):
 
     if settings.USE_DEBUG_STORAGE:
         debug_storage = DebugStorage.instance()
@@ -444,16 +465,35 @@ def test_deploy_and_invoke(deploy_script, invoke_args, wallet,
         invoke_args, neo_to_attach, gas_to_attach = PromptUtils.get_asset_attachments(invoke_args)
         invoke_args, no_parse_addresses = PromptUtils.get_parse_addresses(invoke_args)
 
-        invoke_args.reverse()
-
         if '--i' in invoke_args:
             invoke_args = []
             for index, iarg in enumerate(contract_state.Code.ParameterList):
                 param, abort = PromptUtils.gather_param(index, iarg)
                 if abort:
                     return None, [], 0, None
-                else:
-                    invoke_args.append(param)
+                invoke_args.append(param)
+            invoke_args.reverse()
+        elif user_entry:
+            try:
+                i_args = []
+                for index, iarg in enumerate(contract_state.Code.ParameterList):
+                    ptype = ContractParameterType(iarg)
+                    param, abort = PromptUtils.verify_params(ptype, invoke_args[index])
+                    if abort:
+                        return None, [], 0, None
+                    i_args.append(param)
+                i_args.reverse()
+                invoke_args = i_args
+            except IndexError:
+                print(f"Check inputs. {len(contract_state.Code.ParameterList)} params specified and only {len(invoke_args)} given.")
+                return None, [], 0, None
+            except ValueError as e:
+                print("Check params.", e)
+                return None, [], 0, None
+            except Exception as e:
+                print(f'Could not parse {invoke_args[index]} as {ptype}:', e)
+                return None, [], 0, None
+        else:
             invoke_args.reverse()
 
         sb = ScriptBuilder()