+++ neo/Prompt/Commands/tests/test_sc_commands.py
@@ -128,6 +128,33 @@ class CommandSCTestCase(WalletFixtureTestCase):
             self.assertFalse(tx)
             self.assertIn("run `sc build_run help` to see supported queries", mock_print.getvalue())
 
+        # test too few args
+        PromptData.Wallet = self.GetWallet1(recreate=True)
+        with patch('sys.stdout', new=StringIO()) as mock_print:
+            args = ['build_run', 'neo/Prompt/Commands/tests/SampleSC.py', 'True', 'False', 'False', '070502', '02', 'add', 'AG4GfwjnvydAZodm4xEDivguCtjCFzLcJy',
+                    ]  # missing third param
+            tx, result, total_ops, engine = CommandSC().execute(args)
+            self.assertFalse(tx)
+            self.assertIn("Check inputs. 3 params specified and only 2 given.", mock_print.getvalue())
+
+        # test invalid param type
+        PromptData.Wallet = self.GetWallet1(recreate=True)
+        with patch('sys.stdout', new=StringIO()) as mock_print:
+            args = ['build_run', 'neo/Prompt/Commands/tests/SampleSC.py', 'True', 'False', 'False', 'fe0502', '02', 'add', 'AG4GfwjnvydAZodm4xEDivguCtjCFzLcJy',
+                    '3']  # "fe" is an invalid param type
+            tx, result, total_ops, engine = CommandSC().execute(args)
+            self.assertFalse(tx)
+            self.assertIn("Check params. 254 is not a valid ContractParameterType", mock_print.getvalue())
+
+        # test unknown param type
+        PromptData.Wallet = self.GetWallet1(recreate=True)
+        with patch('sys.stdout', new=StringIO()) as mock_print:
+            args = ['build_run', 'neo/Prompt/Commands/tests/SampleSC.py', 'True', 'False', 'False', '030502', '02', 'add', 'AG4GfwjnvydAZodm4xEDivguCtjCFzLcJy',
+                    '3']  # "03" is Hash160
+            tx, result, total_ops, engine = CommandSC().execute(args)
+            self.assertFalse(tx)
+            self.assertIn('Could not parse add as Hash160: Unknown param type Hash160', mock_print.getvalue())
+
         # test successful build and run
         PromptData.Wallet = self.GetWallet1(recreate=True)
         with patch('sys.stdout', new=StringIO()) as mock_print:
@@ -378,6 +405,13 @@ class CommandSCTestCase(WalletFixtureTestCase):
             res = CommandSC().execute(args)
             self.assertFalse(res)
             self.assertIn("Error testing contract invoke", mock_print.getvalue())
+
+        # test too few args
+        with patch('sys.stdout', new=StringIO()) as mock_print:
+            args = ['invoke', token_hash_str, 'name']  # missing second arg
+            res = CommandSC().execute(args)
+            self.assertFalse(res)
+            self.assertIn("Check inputs. 2 params specified and only 1 given.", mock_print.getvalue())
 
         # test with keyboard interrupt
         with patch('sys.stdout', new=StringIO()) as mock_print: